third_party/WebKit/LayoutTests/vr/getVRDisplays_two_display.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="../resources/mojo-helpers.js"></script>
<script src="resources/fake-vr-displays.js"></script>
<script src="resources/mock-vr-service.js"></script>
<script>
let fakeDisplays = fakeVRDisplays();

vr_test((service) => {
  return navigator.getVRDisplays().then(devices => {
      assert_true(devices != null);
      assert_equals(2, devices.length);
      assert_equals(devices[0].displayName, 'Google, Inc. Daydream View');
      assert_equals(devices[1].displayName, 'FakeVRDisplay');
      assert_true(devices[0].capabilities.hasOrientation);
      assert_true(devices[1].capabilities.hasOrientation);
      assert_true(devices[0].capabilities.canPresent);
      assert_false(devices[1].capabilities.canPresent);
    });
}, [fakeDisplays['Pixel'], fakeDisplays['FakeMagicWindowOnly']],
'Test 2 VRDisplays');

</script>
